services/HealthCheckService/requirements.txt
kafka-python==2.0.2
flask==3.0.3